Fixed those freaking warnings

+++ src/AreaLine.cpp
@@ -21,9 +21,9 @@ void Line::sort() {
     float zA;
     float xB;
     float zB;
-    int size = vertices.size();
-    int i;
-    int j;
+    size_t size = vertices.size();
+    size_t i;
+    size_t j;
     glm::vec4 tmp;
     
     for (i = 0; i < size - 1; i++) {
@@ -45,8 +45,8 @@ void Line::sort() {
 }
 
 AreaLine::AreaLine() {
-    int lineNumber = 0;
-    int i;
+    size_t lineNumber = 0;
+    size_t i;
 
     ifstream arch;
     arch.open("resources/model/area.map");
